requirements.txt
fastapi
uvicorn
python-dotenv
transformers
sentence-transformers
torch
qdrant-client
redis
pydantic
python-multipart